requirements.txt
django~=2.1.0
psycopg2~=2.7.6
scrapy~=1.5.0
python-dateutil~=2.7.0
scrapyd~=1.2.0
python-scrapyd-api~=2.1.0
